Moved the components tests to the components folder
