+++ __tests__/components/ui/Loading.jsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom'
 import { render, screen } from '@testing-library/react'
-import Loading from '../../@core/components/ui/Loading';
+import Loading from '../../../@core/components/ui/Loading';
 
 describe('Loading', () => {
   it('renders a spinner with a "loading" text', () => {
